Moved camera methods to camera service
