+++ TrafficSimulator/TrafficSimulator.Models/Maps/MeterPosition.cs
@@ -7,8 +7,8 @@ namespace TrafficSimulator.Models.Maps
 {
     public class MeterPosition
     {
-        public int X { get; set; };
+        public int X { get; set; }
 
-        public int Y { get; set; };
+        public int Y { get; set; }
     }
 }
